+++ Dockerfile
@@ -2,7 +2,8 @@ FROM golang:latest
 RUN mkdir -p /go/src/app
 WORKDIR /go/src/app
 RUN apt-get -y update && apt-get -y install git
-RUN go get github.com/lib/pq
-RUN go get github.com/google/uuid
-RUN go get github.com/dgrijalva/jwt-go
+COPY app/go.mod .
+COPY app/go.sum .
+RUN go mod download
+COPY . .
 ENTRYPOINT go run main.go
