+++ README.md
@@ -353,3 +353,5 @@
 * как сделать чат.
 * как сделать оповещения.
 
+echo "host    all             all             172.18.0.0/24           md5" >> /var/lib/postgresql/data/pg_hba.conf
+echo "listen_addresses = '*' " >> /var/lib/postgresql/data/postgresql.conf
