+++ app/go.mod
@@ -1,0 +1,9 @@
+module github.com/pristupaanastasia/matcha42/tree/main/app
+
+go 1.15
+
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/google/uuid v1.2.0
+	github.com/lib/pq v1.9.0
+)
